fix(data, data/model): rename data -> power data
