+++ pt/app.py
@@ -20,7 +20,7 @@ def create_app(config_mode):
     def af_request(resp):
         resp = make_response(resp)
         resp.headers['Access-Control-Allow-Origin'] = '*'
-        resp.headers['Access-Control-Allow-Methods'] = 'GET,POST'
+        resp.headers['Access-Control-Allow-Methods'] = 'GET,POST,PUT,DELETE'
         resp.headers['Access-Control-Allow-Headers'] = 'x-requested-with,content-type'
         return resp
