+++ pt/endpoints/__init__.py
@@ -1,7 +1,7 @@
 from endpoints.address.resource import AddressResource, AmiResource
 from endpoints.user.resource import UserResource, LoginResource, ParticipantResource
 from endpoints.news.resource import NewsResource
-from endpoints.data.resource import DatasResource
+from endpoints.powerdata.resource import PowerDatasResource
 from endpoints.bid.resource import BidsResource
 
 
@@ -10,7 +10,7 @@ RESOURCES = {
     'amis': AmiResource,
     'user': UserResource,
     'news': NewsResource,
-    'power_info': DatasResource,
+    'power_info': PowerDatasResource,
     'login': LoginResource,
     'participant': ParticipantResource,
     'bids': BidsResource,
